backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pyodbc==4.0.39

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI and Vector Database
google-generativeai==0.3.2
chromadb==0.4.18
sentence-transformers==2.2.2

# Document processing
pypdf2==3.0.1
python-docx==0.8.11
pandas==2.1.4

# Data validation
pydantic==2.5.0
email-validator==2.1.0

# File handling
aiofiles==23.2.0
